Upgrade to use Lucene 5.4.0
Lucene 5.4.0 still uses the SegmentInfos from Lucene 5.3.x, and so
there's no way to distinguish between 5.3.x and 5.4.x with
SegmentInfos alone.

This updates the package namespace where PayloadNearQuery and
PayloadTermQuery now live.

This also updates the signature of Luke's FsDirectory.obtainLock
to match that of LockFactory's.

Finally, this comes with a few whitespace fixes to match the style of
the neighboring code.